requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.2
azure-storage-blob==12.20.0
azure-storage-file-datalake==12.15.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
bitsandbytes-cuda115==0.26.0.post2
black==24.4.2
bleach==6.1.0
blinker==1.8.2
boto3==1.34.130
botocore==1.34.130
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
coreforecast==0.0.12
cryptography==42.0.8
cycler==0.12.1
databricks==0.2
databricks-cli==0.18.0
databricks-connect==13.3.0
databricks-sdk==0.28.0
datascience_library==2.8.0
datasets==3.0.1
debugpy==1.8.1
decorator==5.1.1
demand-forecast-service @ file:///Users/aifora/Desktop/work/demand_forecast_service/dist/demand_forecast_service-0.0.1-py3-none-any.whl
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
einops==0.8.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.0
fastjsonschema==2.20.0
filelock==3.13.4
Flask==3.0.3
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
groq==0.11.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
holidays==0.46
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.7
img2vec-pytorch==1.0.1
importlib_metadata==7.1.0
interegular==0.3.3
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.5
isodate==0.6.1
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
jupytext==1.16.3
kaggle==1.6.12
kiwisolver==1.4.5
langchain==0.3.1
langchain-core==0.3.6
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.129
lark==1.2.2
lightgbm==4.4.0
lightning-utilities==0.11.6
llvmlite==0.43.0
lm-format-enforcer==0.10.6
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistral_common==1.4.4
mlflow==2.15.1
mlflow-skinny==2.16.2
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
neuralforecast==1.7.4
numba==0.60.0
numpy==1.24.4
nvidia-ml-py==12.560.30
oauthlib==3.2.2
openai==1.45.0
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
optuna==3.6.1
orjson==3.10.7
outlines==0.0.46
p2j==1.3.2
packaging==24.1
pandas==2.2.2
paramiko==2.12.0
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prometheus-fastapi-instrumentator==7.0.0
prometheus-flask-exporter==0.23.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyairports==2.1.1
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.0b1
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
querystring-parser==1.2.4
ray==2.34.0
referencing==0.35.0
regex==2024.4.28
requests==2.32.3
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.9.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.31
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.38.6
statsmodels==0.14.2
streamlit==1.33.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.2
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
utilsforecast==0.2.3
uvicorn==0.31.0
uvloop==0.20.0
vllm==0.6.2
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.3
whittaker-eilers==0.1.3
widgetsnbextension==4.0.13
wrapt==1.16.0
xgboost==2.1.0
xicorpy==0.5
xxhash==3.5.0
yarl==1.9.4
zipp==3.19.2
